# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Uno
  version: "1"
  description: |
    Uno's HTTP surface: the builder API under `/api`, which the SPA calls, and
    the public Stremio-protocol addon under `/u/{token}`, which Nuvio calls.
    Everything else on the server is the embedded SPA and is not described
    here.

    **Errors.** Every error under `/api`, an unknown path included, is JSON
    (`Error`): `error` holds one line of words and, for an error the builder
    acts on by kind, `code` names it (`ProfileNotFound`). Push is the one
    exception: past auth, profile resolution and body decoding it answers
    `PushResult` whether it succeeds or fails. The addon under `/u/{token}`
    answers errors in plain text (`text/plain; charset=utf-8`), which its
    clients don't read.

    **Request bodies.** Every body is decoded strictly: a field the request
    doesn't have, at any depth, is a 400 naming it (`invalid request body:
    unknown field "name"`). The default limit is 1 MiB, past which the answer
    is 413.

    **Auth.** Every `/api` operation except `GET /api/health` and
    `GET /api/config` takes a Nuvio access token as `Authorization: Bearer`.
    Any of them can answer 401 (missing, malformed, invalid or expired
    token), 403 (an account the server's allowlist doesn't list) or 502
    (`auth service unavailable`: Nuvio's signing keys couldn't be fetched).
    Any operation can answer 500 for a failure of Uno's own.

...

...
    TMDB key or TMDB refuses it.

...

...
servers:
...
security:
  - nuvioBearer: []
paths:
...
  /api/genres/{type}:
    parameters:
      - $ref: "#/components/parameters/TypePath"
    get:
      tags: [tmdb]
      operationId: listGenres
      summary: TMDB's genres for a catalog type
      responses:
        "200":
          description: The genres.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Genre"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/certifications/{type}:
    parameters:
      - $ref: "#/components/parameters/TypePath"
    get:
      tags: [tmdb]
      operationId: listCertifications
      summary: TMDB's certifications for a catalog type, by country
      responses:
        "200":
          description: Country code → that country's certifications.
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: "#/components/schemas/Certification"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/watch-providers/{type}:
    parameters:
      - $ref: "#/components/parameters/TypePath"
    get:
      tags: [tmdb]
      operationId: listWatchProviders
      summary: TMDB's streaming services for a catalog type
      parameters:
        - name: watch_region
          in: query
          description: An ISO 3166-1 watch region; omitted, every service TMDB knows.
          schema:
            type: string
      responses:
        "200":
          description: The streaming services.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchProvider"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/languages:
    get:
      tags: [tmdb]
      operationId: listLanguages
      summary: TMDB's languages
      responses:
        "200":
          description: The languages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Language"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/countries:
    get:
      tags: [tmdb]
      operationId: listCountries
      summary: TMDB's countries
      responses:
        "200":
          description: The countries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/watch-regions:
    get:
      tags: [tmdb]
      operationId: listWatchRegions
      summary: TMDB's watch regions
      responses:
        "200":
          description: The watch regions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/companies/search:
    get:
      tags: [tmdb]
      operationId: searchCompanies
      summary: Search TMDB's companies
      description: TMDB's first page of matches, each with its title count for the catalog type.
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: type
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/CatalogType"
      responses:
        "200":
          description: The matches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TitledMatch"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/companies/{id}:
    parameters:
      - $ref: "#/components/parameters/TMDBID"
    get:
      tags: [tmdb]
      operationId: getCompany
      summary: One TMDB company
      responses:
        "200":
          description: The company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Named"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotOnTMDB" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/keywords/search:
    get:
      tags: [tmdb]
      operationId: searchKeywords
      summary: Search TMDB's keywords
      description: TMDB's first page of matches.
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
      responses:
        "200":
          description: The matches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Named"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/keywords/{id}:
    parameters:
      - $ref: "#/components/parameters/TMDBID"
    get:
      tags: [tmdb]
      operationId: getKeyword
      summary: One TMDB keyword
      responses:
        "200":
          description: The keyword.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Named"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotOnTMDB" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/collections/search:
    get:
      tags: [tmdb]
      operationId: searchCollections
      summary: Search TMDB's collections
      description: TMDB's movie collections (e.g. a film series), not Uno collections; first page of matches.
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
      responses:
        "200":
          description: The matches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Named"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/collections/{id}:
    parameters:
      - $ref: "#/components/parameters/TMDBID"
    get:
      tags: [tmdb]
      operationId: getCollection
      summary: One TMDB collection
      responses:
        "200":
          description: The collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Named"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotOnTMDB" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/networks/search:
    get:
      tags: [tmdb]
      operationId: searchNetworks
      summary: Search TV networks
      description: |
        TMDB has no network search: this matches TMDB's daily network export,
        then counts each match's series. Networks filter series only.
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
      responses:
        "200":
          description: The matches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TitledMatch"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/networks/{id}:
    parameters:
      - $ref: "#/components/parameters/TMDBID"
    get:
      tags: [tmdb]
      operationId: getNetwork
      summary: One TV network
      responses:
        "200":
          description: The network.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotOnTMDB" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
...
components:
  securitySchemes:
    nuvioBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Nuvio access token, verified against Nuvio's signing keys.
  parameters:
...
    TypePath:
      name: type
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/CatalogType"
    TMDBID:
      name: id
      in: path
      required: true
      description: A TMDB id.
      schema:
        type: integer
    SearchQuery:
      name: q
      in: query
      required: true
      description: The search text; blank is a 400.
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: The request is malformed or refused in words the caller can act on.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: Missing, malformed, invalid or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The server's allowlist doesn't list this Nuvio account.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotOnTMDB:
      description: TMDB has no such id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    KeyProblem:
      description: Per-account key server only. The account has no TMDB key yet, or TMDB doesn't accept it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TMDBUnavailable:
      description: TMDB couldn't be reached, or the auth service is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    CatalogType:
      type: string
      enum: [movie, series]
...
    Error:
      type: object
      description: The answer to every failed request under `/api` but push's own.
      required: [error]
      properties:
        error:
          type: string
          description: One line of words, for a person to read.
        code:
          type: string
          description: A stable name for an error the builder acts on by kind rather than by status; absent when there is none.
...
    Genre:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
    Named:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
    TitledMatch:
      type: object
      required: [id, name, origin_country, title_count]
      properties:
        id: { type: integer }
        name: { type: string }
        origin_country: { type: string }
        title_count:
          type: integer
          description: How many titles of the requested type it has (series, for a network).
    Network:
      type: object
      required: [id, name, origin_country]
      properties:
        id: { type: integer }
        name: { type: string }
        origin_country: { type: string }
    Language:
      type: object
      required: [iso_639_1, english_name, name]
      properties:
        iso_639_1: { type: string }
        english_name: { type: string }
        name: { type: string }
    Country:
      type: object
      required: [iso_3166_1, english_name, native_name]
      properties:
        iso_3166_1: { type: string }
        english_name: { type: string }
        native_name: { type: string }
    WatchProvider:
      type: object
      required: [provider_id, provider_name, display_priority, logo_path]
      properties:
        provider_id: { type: integer }
        provider_name: { type: string }
        display_priority: { type: integer }
        logo_path: { type: string }
    Certification:
      type: object
      required: [certification, meaning, order]
      properties:
        certification: { type: string }
        meaning: { type: string }
        order: { type: integer }
...